packet descriptor is 64 bits

+++ sw/pionic.c
@@ -105,7 +105,7 @@ void pionic_fini(pionic_ctx_t *ctx) {
 bool pionic_rx(pionic_ctx_t *ctx, int cid, pionic_pkt_desc_t *desc) {
   uint64_t reg = read64(ctx, PIONIC_CONTROL_HOST_RX_NEXT(cid));
   if (reg & 0x1) {
-    uint32_t hw_desc = reg >> 1;
+    uint64_t hw_desc = reg >> 1;
     uint64_t read_addr = PIONIC_PKTBUF_OFF_TO_ADDR(hw_desc & PIONIC_PKT_ADDR_MASK);
     uint64_t pkt_len = (hw_desc >> PIONIC_PKT_ADDR_WIDTH) & PIONIC_PKT_LEN_MASK;
     desc->buf = (uint8_t *)(ctx->bar) + read_addr;
@@ -131,7 +131,7 @@ void pionic_rx_ack(pionic_ctx_t *ctx, int cid, pionic_pkt_desc_t *desc) {
 
 void pionic_tx_get_desc(pionic_ctx_t *ctx, int cid, pionic_pkt_desc_t *desc) {
   uint64_t reg = read64(ctx, PIONIC_CONTROL_HOST_TX(cid));
-  uint32_t hw_desc = reg >> 1;
+  uint64_t hw_desc = reg >> 1;
 
   desc->buf = (uint8_t *)(ctx->bar) + PIONIC_PKTBUF_OFF_TO_ADDR(hw_desc & PIONIC_PKT_ADDR_MASK);
   desc->len = (hw_desc >> PIONIC_PKT_ADDR_WIDTH) & PIONIC_PKT_LEN_MASK;
